Add parameterized resolve/state test
